requirements.txt
# --- GUI / Qt ---
PySide6>=6.6

# --- Computer Vision ---
opencv-python>=4.8
numpy>=1.26
Pillow>=10.0

# --- IA / Deep Learning ---
torch>=2.1
ultralytics>=8.0

# --- Scientific / Math ---
scipy>=1.11
networkx>=3.2

# --- Visualization ---
matplotlib>=3.8

# --- PDF / Reports ---
reportlab>=4.0

# --- Theming ---
qdarktheme>=2.1
darkdetect>=0.8

# --- Web / Mobile ---
Flask>=3.0
qrcode>=7.4

# --- System / Windows ---
pywin32>=306
winshell>=0.6
psutil>=5.9

# --- GPU (optional, safe fallback) ---
pynvml>=11.5

# --- Utilities ---
requests>=2.31
